Improved GSS need for rekey detection.
Use time_t for time, make clear when we're dealing with absolute time
vs. relative time, and add support for DIR ccaches (effectively).

The updated code now saves the previous TGT expiration and compares
that to the new TGT expiration.  In addition it rekeys shortly before
the service ticket expiration time provided we have enough time left in
the TGT (because the delegated TGT's lifetime may be that of the service
ticket rather than the TGT).

To obtain the TGT expiration, acquire a credential for the saved name,
and then call gss_inquire_cred_by_mech() on that.

To obtain the service ticket expiration we use gss_inquire_context().
We recover from any (unlikely) failure of gss_inquire_context(), by using
the default name and mechanism until it does succeed on some future rekey.
Under the same conditions we guess the expiration time to be 1 hour in
the future.