Allow GSSAPI work to localhost by using actual hostname
This feature is off by default, enabled via the ReplaceLocalhost boolean
ssh_config(5) option.

If the connection is not proxied and the given host's IP address(es) are
all loopback addresses, then replace the given hostname with "thishost"
(the local host's hostname).

Similar logic is possible via the new %l and %L token expansions for
for HostName option.