Overhaul GSSAPI kex state.
Moved saved expirations, mech oid and client name into the kex->gss
structure.  When rekeying with GSSAPI only use the initial mechanism.

Now that Heimdal's gss_acquire_cred() with an explicit name no longer
fails to obtain the expiration time, use the saved name to test and
delegate the right credential.  In particular if the user switches to
a different identity we avoid GSS key exchange (and delegation of the
wrong credentials).